+++ bank-feeds/CodatBankFeeds/CodatBankFeedsSDK.cs
@@ -18,7 +18,7 @@ namespace CodatBankFeeds
     using System;
 
     /// <summary>
-    /// Bank Feeds API: Bank Feeds API enables your SMB users to set up bank feeds from accounts in your application to supported accounting platforms.&lt;br/&gt;
+    /// Bank Feeds API: Bank Feeds API enables your SMB users to set up bank feeds from accounts in your application to supported accounting platforms.<br/>
     /// 
     /// <remarks>
     /// <br/>
@@ -53,7 +53,7 @@ namespace CodatBankFeeds
         public ICompaniesSDK Companies { get; }
 
         /// <summary>
-        /// Manage your companies&amp;apos; data connections.
+        /// Manage your companies&apos; data connections.
         /// </summary>
         public IConnectionsSDK Connections { get; }
 
@@ -73,7 +73,7 @@ namespace CodatBankFeeds
     }
 
     /// <summary>
-    /// Bank Feeds API: Bank Feeds API enables your SMB users to set up bank feeds from accounts in your application to supported accounting platforms.&lt;br/&gt;
+    /// Bank Feeds API: Bank Feeds API enables your SMB users to set up bank feeds from accounts in your application to supported accounting platforms.<br/>
     /// 
     /// <remarks>
     /// <br/>
@@ -103,9 +103,10 @@ namespace CodatBankFeeds
         };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "2.1.0";
-        private const string _sdkGenVersion = "2.129.1";
+        private const string _sdkVersion = "2.2.0";
+        private const string _sdkGenVersion = "2.144.7";
         private const string _openapiDocVersion = "3.0.0";
+        private const string _userAgent = "speakeasy-sdk/csharp 2.2.0 2.144.7 3.0.0 Codat.BankFeeds";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;