+++ bank-feeds/RELEASES.md
@@ -89,3 +89,13 @@ Based on:
 - [csharp v2.1.0] bank-feeds
 ### Releases
 - [NuGet v2.1.0] https://www.nuget.org/packages/Codat.BankFeeds/2.1.0 - bank-feeds
+
+## 2023-10-04 00:12:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.2.0] bank-feeds
+### Releases
+- [NuGet v2.2.0] https://www.nuget.org/packages/Codat.BankFeeds/2.2.0 - bank-feeds